Move settings to its own module.
